pipeline/data_retrieval/requirements.txt
kafka-python
pyspark
pandas
schedule
requests
numpy
python-dotenv